Ability to reset the driver
Ugly, but necessary to work with tasks/workers: the notifier is loaded
and initialized before the environment in which will run the tasks is
set, so the notifier load only the "all" keys.

resetDriver() allows to reload a fresh driver, after driver options have
been overrided in the task. Example:

    public function execute($arguments, $options)
    {
      if ($options['env'] == 'prod) {
        sfConfig::set('sf_notify_driver', array(
          'class'   => ....,,
          'options' => array( .... )
        );

        fpErrorNotifier::getInstance()->resetDriver();

      ...
   }